Move external servers related classes to separated package
