+++ infrastructures/kubernetes/src/main/java/org/eclipse/che/workspace/infrastructure/kubernetes/server/external/SingleHostIngressExternalServerExposer.java
@@ -8,36 +8,34 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.workspace.infrastructure.kubernetes.server;
+package org.eclipse.che.workspace.infrastructure.kubernetes.server.external;
 
-import static java.lang.String.format;
-import static org.eclipse.che.workspace.infrastructure.kubernetes.server.ExternalServerExposerStrategyProvider.STRATEGY_PROPERTY;
-
-import com.google.common.base.Strings;
 import io.fabric8.kubernetes.api.model.ServicePort;
 import io.fabric8.kubernetes.api.model.extensions.Ingress;
 import java.util.Map;
 import javax.inject.Inject;
 import javax.inject.Named;
 import org.eclipse.che.api.core.model.workspace.config.ServerConfig;
-import org.eclipse.che.inject.ConfigurationException;
 
 /**
- * Provides a host-based strategy for exposing service ports outside the cluster using Ingress
+ * Provides a path-based strategy for exposing service ports outside the cluster using Ingress.
+ * Ingresses will be created with a common host name for all workspaces.
  *
- * <p>This strategy uses different Ingress host entries <br>
- * Each external server is exposed with a unique subdomain of CHE_DOMAIN.
+ * <p>This strategy uses different Ingress path entries <br>
+ * Each external server is exposed with a unique path prefix.
+ *
+ * <p>This strategy imposes limitation on user-developed applications. <br>
  *
  * <pre>
- *   Host-Based Ingress exposing service's port:
+ *   Path-Based Ingress exposing service's port:
  * Ingress
  * ...
  * spec:
  *   rules:
- *     - host: service123-webapp.che-domain   ---->> Service.metadata.name + - + Service.spec.ports[0].name + . + CHE_DOMAIN
- *     - http:
+ *     - host: CHE_HOST
+ *       http:
  *         paths:
- *           - path: /
+ *           - path: service123/webapp        ---->> Service.metadata.name + / + Service.spec.ports[0].name
  *             backend:
  *               serviceName: service123      ---->> Service.metadata.name
  *               servicePort: [8080|web-app]  ---->> Service.spec.ports[0].[port|name]
@@ -46,29 +44,19 @@ import org.eclipse.che.inject.ConfigurationException;
  * @author Sergii Leshchenko
  * @author Guy Daich
  */
-public class MultiHostIngressExternalServerExposer
+public class SingleHostIngressExternalServerExposer
     extends AbstractIngressExternalServerExposerStrategy {
 
-  public static final String MULTI_HOST_STRATEGY = "multi-host";
-  private static final String INGRESS_DOMAIN_PROPERTY = "che.infra.kubernetes.ingress.domain";
-
-  private final String domain;
+  public static final String SINGLE_HOST_STRATEGY = "single-host";
   private final Map<String, String> ingressAnnotations;
+  private final String cheHost;
 
   @Inject
-  public MultiHostIngressExternalServerExposer(
+  public SingleHostIngressExternalServerExposer(
       @Named("infra.kubernetes.ingress.annotations") Map<String, String> ingressAnnotations,
-      @Named(INGRESS_DOMAIN_PROPERTY) String domain,
-      @Named(STRATEGY_PROPERTY) String strategy) {
-    if (Strings.isNullOrEmpty(domain) && MULTI_HOST_STRATEGY.equals(strategy)) {
-      throw new ConfigurationException(
-          format(
-              "Strategy of generating ingress URLs for Che servers is set to '%s', "
-                  + "but property '%s' is not set",
-              MULTI_HOST_STRATEGY, INGRESS_DOMAIN_PROPERTY));
-    }
+      @Named("che.host") String cheHost) {
     this.ingressAnnotations = ingressAnnotations;
-    this.domain = domain;
+    this.cheHost = cheHost;
   }
 
   @Override
@@ -78,8 +66,8 @@ public class MultiHostIngressExternalServerExposer
       ServicePort servicePort,
       Map<String, ServerConfig> ingressesServers) {
     return new ExternalServerIngressBuilder()
-        .withHost(generateExternalServerIngressHostname(serviceName, servicePort))
-        .withPath(generateExternalServerIngressPath())
+        .withHost(cheHost)
+        .withPath(generateExternalServerIngressPath(serviceName, servicePort))
         .withName(generateExternalServerIngressName(serviceName, servicePort))
         .withMachineName(machineName)
         .withServiceName(serviceName)
@@ -89,16 +77,11 @@ public class MultiHostIngressExternalServerExposer
         .build();
   }
 
-  private String generateExternalServerIngressPath() {
-    return "/";
-  }
-
   private String generateExternalServerIngressName(String serviceName, ServicePort servicePort) {
     return serviceName + '-' + servicePort.getName();
   }
 
-  private String generateExternalServerIngressHostname(
-      String serviceName, ServicePort servicePort) {
-    return serviceName + "-" + servicePort.getName() + "." + domain;
+  private String generateExternalServerIngressPath(String serviceName, ServicePort servicePort) {
+    return "/" + serviceName + "/" + servicePort.getName();
   }
 }
