+++ infrastructures/kubernetes/src/test/java/org/eclipse/che/workspace/infrastructure/kubernetes/server/external/DefaultHostIngressExternalServerExposerTest.java
@@ -8,13 +8,12 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.workspace.infrastructure.kubernetes.server;
+package org.eclipse.che.workspace.infrastructure.kubernetes.server.external;
 
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonMap;
 import static org.eclipse.che.workspace.infrastructure.kubernetes.server.KubernetesServerExposer.SERVER_PREFIX;
-import static org.eclipse.che.workspace.infrastructure.kubernetes.server.MultiHostIngressExternalServerExposer.MULTI_HOST_STRATEGY;
-import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.*;
 
 import com.google.common.collect.ImmutableMap;
 import io.fabric8.kubernetes.api.model.Container;
@@ -36,14 +35,13 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 /** @author Guy Daich */
-public class MultiHostIngressExternalServerExposerTest {
+public class DefaultHostIngressExternalServerExposerTest {
 
   private static final Map<String, String> ATTRIBUTES_MAP = singletonMap("key", "value");
   private static final String MACHINE_NAME = "pod/main";
   private static final String SERVICE_NAME = SERVER_PREFIX + "12345678" + "-" + MACHINE_NAME;
-  private static final String DOMAIN = "che.com";
 
-  private MultiHostIngressExternalServerExposer externalServerExposer;
+  private DefaultHostIngressExternalServerExposer externalServerExposer;
   private KubernetesEnvironment kubernetesEnvironment;
 
   @BeforeMethod
@@ -61,8 +59,7 @@ public class MultiHostIngressExternalServerExposerTest {
 
     kubernetesEnvironment =
         KubernetesEnvironment.builder().setPods(ImmutableMap.of("pod", pod)).build();
-    externalServerExposer =
-        new MultiHostIngressExternalServerExposer(emptyMap(), DOMAIN, MULTI_HOST_STRATEGY);
+    externalServerExposer = new DefaultHostIngressExternalServerExposer(emptyMap());
   }
 
   @Test
@@ -70,13 +67,12 @@ public class MultiHostIngressExternalServerExposerTest {
     // given
     ServerConfigImpl httpServerConfig =
         new ServerConfigImpl("8080/tcp", "http", "/api", ATTRIBUTES_MAP);
-    IntOrString targetPort = new IntOrString(8080);
     ServicePort servicePort =
         new ServicePortBuilder()
             .withName("server-8080")
             .withPort(8080)
             .withProtocol("TCP")
-            .withTargetPort(targetPort)
+            .withTargetPort(new IntOrString(8080))
             .build();
     Map<String, ServicePort> portToServicePort = ImmutableMap.of("8080/tcp", servicePort);
     Map<String, ServerConfig> serversToExpose = ImmutableMap.of("http-server", httpServerConfig);
@@ -103,14 +99,12 @@ public class MultiHostIngressExternalServerExposerTest {
         new ServerConfigImpl("8080/tcp", "http", "/api", ATTRIBUTES_MAP);
     ServerConfigImpl wsServerConfig =
         new ServerConfigImpl("8080/tcp", "ws", "/connect", ATTRIBUTES_MAP);
-    IntOrString targetPort = new IntOrString(8080);
-
     ServicePort servicePort =
         new ServicePortBuilder()
             .withName("server-8080")
             .withPort(8080)
             .withProtocol("TCP")
-            .withTargetPort(targetPort)
+            .withTargetPort(new IntOrString(8080))
             .build();
     Map<String, ServicePort> portToServicePort = ImmutableMap.of("8080/tcp", servicePort);
 
@@ -150,21 +144,19 @@ public class MultiHostIngressExternalServerExposerTest {
         new ServerConfigImpl("8080/tcp", "http", "/api", ATTRIBUTES_MAP);
     ServerConfigImpl wsServerConfig =
         new ServerConfigImpl("8081/tcp", "ws", "/connect", ATTRIBUTES_MAP);
-    IntOrString httpTargetPort = new IntOrString(8080);
-    IntOrString wsTargetPort = new IntOrString(8081);
     ServicePort httpServicePort =
         new ServicePortBuilder()
             .withName("server-8080")
             .withPort(8080)
             .withProtocol("TCP")
-            .withTargetPort(httpTargetPort)
+            .withTargetPort(new IntOrString(8080))
             .build();
     ServicePort wsServicePort =
         new ServicePortBuilder()
             .withName("server-8081")
             .withPort(8081)
             .withProtocol("TCP")
-            .withTargetPort(wsTargetPort)
+            .withTargetPort(new IntOrString(8081))
             .build();
     Map<String, ServicePort> portToServicePort =
         ImmutableMap.of("8080/tcp", httpServicePort, "8081/tcp", wsServicePort);
@@ -210,8 +202,6 @@ public class MultiHostIngressExternalServerExposerTest {
     // ensure that required ingress is created
     Ingress ingress = kubernetesEnvironment.getIngresses().get(serviceName + "-server-" + port);
     IngressRule ingressRule = ingress.getSpec().getRules().get(0);
-    assertEquals(ingressRule.getHost(), serviceName + "-" + servicePort.getName() + "." + DOMAIN);
-    assertEquals(ingressRule.getHttp().getPaths().get(0).getPath(), "/");
     IngressBackend backend = ingressRule.getHttp().getPaths().get(0).getBackend();
     assertEquals(backend.getServiceName(), serviceName);
     assertEquals(backend.getServicePort().getStrVal(), servicePort.getName());
